+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - circularAssetImage, circularNetworImage, circularAssetShadowImage added to string extension methods.
 - orientation prop added to context extensions.
 - randomColor, randomOpaqueColor and randomPrimaryColor added to Vx.
+- VxPopupMenu widget Added.
 - Added timeAgo ext for DateTime.
 - [BREAKING] Overflow has been removed and clip has been added with default as Clipbehavior.None. It can affect all the widgets which uses Stack or ZStack.